PMM-6469 Refactor process promise results to shared
