+++ pmm-app/src/pmm-inventory/Tabs/Agents.tsx
@@ -3,7 +3,8 @@ import { Button, HorizontalGroup, Modal } from '@grafana/ui';
 import { Form } from 'react-final-form';
 import { Table } from 'shared/components/Elements/Table/Table';
 import { showSuccessNotification } from 'shared/components/helpers';
-import { filterFulfilled, InventoryDataService, processPromiseResults } from 'pmm-inventory/Inventory.tools';
+import { filterFulfilled, processPromiseResults } from 'shared/components/helpers/promises';
+import { InventoryDataService } from 'pmm-inventory/Inventory.tools';
 import { CheckboxField, FormElement } from 'shared/components/Form';
 import { AgentsList } from 'pmm-inventory/Inventory.types';
 import { SelectedTableRows } from 'shared/components/Elements/Table/Table.types';
