+++ pmm-app/src/pmm-inventory/Tabs/Nodes.tsx
@@ -4,7 +4,8 @@ import { Form } from 'react-final-form';
 import { Table } from 'shared/components/Elements/Table/Table';
 import { showSuccessNotification } from 'shared/components/helpers';
 import { CheckboxField, FormElement } from 'shared/components/Form';
-import { filterFulfilled, InventoryDataService, processPromiseResults } from 'pmm-inventory/Inventory.tools';
+import { filterFulfilled, processPromiseResults } from 'shared/components/helpers/promises';
+import { InventoryDataService } from 'pmm-inventory/Inventory.tools';
 import { SelectedTableRows } from 'shared/components/Elements/Table/Table.types';
 import { InventoryService } from '../Inventory.service';
 import { NodesList } from '../Inventory.types';
